chore(account): add account module and fix request to be signed in the request method
