+++ .byebug_history
@@ -1,0 +1,16 @@
+c
+request
+c
+exit
+request
+connection
+request.headers['potato'] = 'adfadfad'
+request.headers
+request
+c
+exit
+s
+c
+connection
+connection.headers['Apiauth-Potato'] = 'adfadf'
+connection.headers
