+++ lib/localbitcoins/endpoint.rb
@@ -1,4 +1,7 @@
+require './lib/localbitcoins/endpoint/account'
+
 module Localbitcoins
   module Endpoint
+    include Account
   end
 end
